Move long tasks utils to utils folder
